feat - 채팅 내 up 버튼 추가 - 채팅 input 분리 - ChatMain -> ChatContent로 변경
